move rulesDeploy file
